factored out bulkop, insert, find, and raw into common.js
