+++ lib/bulk/common.js
@@ -5,6 +5,9 @@ const MongoError = require('mongodb-core').MongoError;
 const util = require('util');
 const toError = require('../utils').toError;
 const handleCallback = require('../utils').handleCallback;
+const applyWriteConcern = require('../utils').applyWriteConcern;
+const shallowClone = require('../utils').shallowClone;
+const ObjectID = require('mongodb-core').BSON.ObjectID;
 
 // Error codes
 var UNKNOWN_ERROR = 8;
@@ -503,15 +506,209 @@ const Delete = function(bulk) {
   bulk.s.currentOp = null;
   return bulk.addToOperationsList(bulk, REMOVE, document);
 };
+
+const BulkOperation = function(topology, collection, options, bulk) {
+  options = options == null ? {} : options;
+  // TODO Bring from driver information in isMaster
+  // Get the namespace for the write operations
+  var namespace = collection.collectionName;
+  // Used to mark operation as executed
+  var executed = false;
+
+  // Current item
+  var currentOp = null;
+
+  // Handle to the bson serializer, used to calculate running sizes
+  var bson = topology.bson;
+
+  // Set max byte size
+  var maxBatchSizeBytes =
+    topology.isMasterDoc && topology.isMasterDoc.maxBsonObjectSize
+      ? topology.isMasterDoc.maxBsonObjectSize
+      : 1024 * 1025 * 16;
+  var maxWriteBatchSize =
+    topology.isMasterDoc && topology.isMasterDoc.maxWriteBatchSize
+      ? topology.isMasterDoc.maxWriteBatchSize
+      : 1000;
+
+  // Get the write concern
+  var writeConcern = applyWriteConcern(shallowClone(options), { collection: collection }, options);
+  writeConcern = writeConcern.writeConcern;
+
+  // Get the promiseLibrary
+  var promiseLibrary = options.promiseLibrary || Promise;
+
+  // Final results
+  var bulkResult = {
+    ok: 1,
+    writeErrors: [],
+    writeConcernErrors: [],
+    insertedIds: [],
+    nInserted: 0,
+    nUpserted: 0,
+    nMatched: 0,
+    nModified: 0,
+    nRemoved: 0,
+    upserted: []
+  };
+
+  // Internal state
+  bulk.s = {
+    // Final result
+    bulkResult: bulkResult,
+    // Current batch state
+    currentBatch: null,
+    currentIndex: 0,
+    // ordered specific
+    currentBatchSize: 0,
+    currentBatchSizeBytes: 0,
+    // unordered specific
+    currentInsertBatch: null,
+    currentUpdateBatch: null,
+    currentRemoveBatch: null,
+    batches: [],
+    // Write concern
+    writeConcern: writeConcern,
+    // Max batch size options
+    maxBatchSizeBytes: maxBatchSizeBytes,
+    maxWriteBatchSize: maxWriteBatchSize,
+    // Namespace
+    namespace: namespace,
+    // BSON
+    bson: bson,
+    // Topology
+    topology: topology,
+    // Options
+    options: options,
+    // Current operation
+    currentOp: currentOp,
+    // Executed
+    executed: executed,
+    // Collection
+    collection: collection,
+    // Promise Library
+    promiseLibrary: promiseLibrary,
+    // Fundamental error
+    err: null,
+    // check keys
+    checkKeys: typeof options.checkKeys === 'boolean' ? options.checkKeys : true
+  };
+
+  // bypass Validation
+  if (options.bypassDocumentValidation === true) {
+    bulk.s.bypassDocumentValidation = true;
+  }
+};
+
+const Insert = function(document, bulk) {
+  if (bulk.s.collection.s.db.options.forceServerObjectId !== true && document._id == null)
+    document._id = new ObjectID();
+  return bulk.addToOperationsList(bulk, INSERT, document);
+};
+
+const Find = function(selector, bulk) {
+  if (!selector) {
+    throw toError('Bulk find operation must specify a selector');
+  }
+
+  // Save a current selector
+  bulk.s.currentOp = {
+    selector: selector
+  };
+};
+
+const Raw = function(op, bulk) {
+  var key = Object.keys(op)[0];
+
+  // Set up the force server object id
+  var forceServerObjectId =
+    typeof bulk.s.options.forceServerObjectId === 'boolean'
+      ? bulk.s.options.forceServerObjectId
+      : bulk.s.collection.s.db.options.forceServerObjectId;
+
+  // Update operations
+  if (
+    (op.updateOne && op.updateOne.q) ||
+    (op.updateMany && op.updateMany.q) ||
+    (op.replaceOne && op.replaceOne.q)
+  ) {
+    op[key].multi = op.updateOne || op.replaceOne ? false : true;
+    return bulk.addToOperationsList(bulk, UPDATE, op[key]);
+  }
+
+  // Crud spec update format
+  if (op.updateOne || op.updateMany || op.replaceOne) {
+    var multi = op.updateOne || op.replaceOne ? false : true;
+    var operation = { q: op[key].filter, u: op[key].update || op[key].replacement, multi: multi };
+    if (bulk.isOrdered()) {
+      operation.upsert = op[key].upsert ? true : false;
+      if (op.collation) operation.collation = op.collation;
+    } else {
+      if (op[key].upsert) operation.upsert = true;
+    }
+    if (op[key].arrayFilters) operation.arrayFilters = op[key].arrayFilters;
+    return bulk.addToOperationsList(bulk, UPDATE, operation);
+  }
+
+  // Remove operations
+  if (
+    op.removeOne ||
+    op.removeMany ||
+    (op.deleteOne && op.deleteOne.q) ||
+    (op.deleteMany && op.deleteMany.q)
+  ) {
+    op[key].limit = op.removeOne ? 1 : 0;
+    return bulk.addToOperationsList(bulk, REMOVE, op[key]);
+  }
+
+  // Crud spec delete operations, less efficient
+  if (op.deleteOne || op.deleteMany) {
+    var limit = op.deleteOne ? 1 : 0;
+    operation = { q: op[key].filter, limit: limit };
+    if (bulk.isOrdered()) {
+      if (op.collation) operation.collation = op.collation;
+    }
+    return bulk.addToOperationsList(bulk, REMOVE, operation);
+  }
+
+  // Insert operations
+  if (op.insertOne && op.insertOne.document == null) {
+    if (forceServerObjectId !== true && op.insertOne._id == null) op.insertOne._id = new ObjectID();
+    return bulk.addToOperationsList(bulk, INSERT, op.insertOne);
+  } else if (op.insertOne && op.insertOne.document) {
+    if (forceServerObjectId !== true && op.insertOne.document._id == null)
+      op.insertOne.document._id = new ObjectID();
+    return bulk.addToOperationsList(bulk, INSERT, op.insertOne.document);
+  }
+
+  if (op.insertMany) {
+    for (var i = 0; i < op.insertMany.length; i++) {
+      if (forceServerObjectId !== true && op.insertMany[i]._id == null)
+        op.insertMany[i]._id = new ObjectID();
+      bulk.addToOperationsList(bulk, INSERT, op.insertMany[i]);
+    }
+
+    return;
+  }
+
+  // No valid type of operation
+  throw toError(
+    'bulkWrite only supports insertOne, insertMany, updateOne, updateMany, removeOne, removeMany, deleteOne, deleteMany'
+  );
+};
+
 // Exports symbols
 module.exports = {
   Batch,
+  BulkOperation,
   BulkWriteError,
   BulkWriteResult,
   cloneOptions,
   Delete,
   DeleteOne,
+  Find,
   handleMongoWriteConcernError,
+  Insert,
   LegacyOp,
   mergeBatchResults,
   INVALID_BSON_ERROR: INVALID_BSON_ERROR,
@@ -521,6 +718,7 @@ module.exports = {
   INSERT: INSERT,
   UPDATE: UPDATE,
   REMOVE: REMOVE,
+  Raw,
   WriteError,
   WriteConcernError,
   Update,
