+++ lib/bulk/ordered.js
@@ -4,15 +4,12 @@ const common = require('./common');
 const utils = require('../utils');
 const toError = require('../utils').toError;
 const handleCallback = require('../utils').handleCallback;
-const shallowClone = utils.shallowClone;
 const BulkWriteResult = common.BulkWriteResult;
-const ObjectID = require('mongodb-core').BSON.ObjectID;
 const BSON = require('mongodb-core').BSON;
 const Batch = common.Batch;
 const mergeBatchResults = common.mergeBatchResults;
 const executeOperation = utils.executeOperation;
 const BulkWriteError = require('./common').BulkWriteError;
-const applyWriteConcern = utils.applyWriteConcern;
 const MongoWriteConcernError = require('mongodb-core').MongoWriteConcernError;
 const handleMongoWriteConcernError = require('./common').handleMongoWriteConcernError;
 
@@ -177,104 +174,19 @@ var addToOperationsList = function(_self, docType, document) {
  * @property {number} length Get the number of operations in the bulk.
  * @return {OrderedBulkOperation} a OrderedBulkOperation instance.
  */
-function OrderedBulkOperation(topology, collection, options) {
-  options = options == null ? {} : options;
-  // TODO Bring from driver information in isMaster
-  var executed = false;
-
-  // Current item
-  var currentOp = null;
-
-  // Handle to the bson serializer, used to calculate running sizes
-  var bson = topology.bson;
-
-  // Namespace for the operation
-  var namespace = collection.collectionName;
-
-  // Set max byte size
-  var maxBatchSizeBytes =
-    topology.isMasterDoc && topology.isMasterDoc.maxBsonObjectSize
-      ? topology.isMasterDoc.maxBsonObjectSize
-      : 1024 * 1025 * 16;
-  var maxWriteBatchSize =
-    topology.isMasterDoc && topology.isMasterDoc.maxWriteBatchSize
-      ? topology.isMasterDoc.maxWriteBatchSize
-      : 1000;
-
-  // Get the write concern
-  var writeConcern = applyWriteConcern(shallowClone(options), { collection: collection }, options);
-  writeConcern = writeConcern.writeConcern;
-
-  // Get the promiseLibrary
-  var promiseLibrary = options.promiseLibrary || Promise;
-
-  // Final results
-  var bulkResult = {
-    ok: 1,
-    writeErrors: [],
-    writeConcernErrors: [],
-    insertedIds: [],
-    nInserted: 0,
-    nUpserted: 0,
-    nMatched: 0,
-    nModified: 0,
-    nRemoved: 0,
-    upserted: []
-  };
-
-  // Internal state
-  this.s = {
-    // Final result
-    bulkResult: bulkResult,
-    // Current batch state
-    currentBatch: null,
-    currentIndex: 0,
-    currentBatchSize: 0,
-    currentBatchSizeBytes: 0,
-    batches: [],
-    // Write concern
-    writeConcern: writeConcern,
-    // Max batch size options
-    maxBatchSizeBytes: maxBatchSizeBytes,
-    maxWriteBatchSize: maxWriteBatchSize,
-    // Namespace
-    namespace: namespace,
-    // BSON
-    bson: bson,
-    // Topology
-    topology: topology,
-    // Options
-    options: options,
-    // Current operation
-    currentOp: currentOp,
-    // Executed
-    executed: executed,
-    // Collection
-    collection: collection,
-    // Promise Library
-    promiseLibrary: promiseLibrary,
-    // Fundamental error
-    err: null,
-    // check keys
-    checkKeys: typeof options.checkKeys === 'boolean' ? options.checkKeys : true
-  };
-  // bypass Validation
-  if (options.bypassDocumentValidation === true) {
-    this.s.bypassDocumentValidation = true;
-  }
-}
+const OrderedBulkOperation = function(topology, collection, options) {
+  common.BulkOperation(topology, collection, options, this);
+};
 
 /**
  * Add a single insert document to the bulk operation
  *
- * @param {object} doc the document to insert
+ * @param {object} document the document to insert
  * @throws {MongoError}
  * @return {OrderedBulkOperation}
  */
 OrderedBulkOperation.prototype.insert = function(document) {
-  if (this.s.collection.s.db.options.forceServerObjectId !== true && document._id == null)
-    document._id = new ObjectID();
-  return this.addToOperationsList(this, common.INSERT, document);
+  return common.Insert(document, this);
 };
 
 /**
@@ -286,15 +198,7 @@ OrderedBulkOperation.prototype.insert = function(document) {
  * @return {FindOperatorsOrdered}
  */
 OrderedBulkOperation.prototype.find = function(selector) {
-  if (!selector) {
-    throw toError('Bulk find operation must specify a selector');
-  }
-
-  // Save a current selector
-  this.s.currentOp = {
-    selector: selector
-  };
-
+  common.Find(selector, this);
   return new FindOperatorsOrdered(this);
 };
 
@@ -306,77 +210,7 @@ Object.defineProperty(OrderedBulkOperation.prototype, 'length', {
 });
 
 OrderedBulkOperation.prototype.raw = function(op) {
-  var key = Object.keys(op)[0];
-
-  // Set up the force server object id
-  var forceServerObjectId =
-    typeof this.s.options.forceServerObjectId === 'boolean'
-      ? this.s.options.forceServerObjectId
-      : this.s.collection.s.db.options.forceServerObjectId;
-
-  // Update operations
-  if (
-    (op.updateOne && op.updateOne.q) ||
-    (op.updateMany && op.updateMany.q) ||
-    (op.replaceOne && op.replaceOne.q)
-  ) {
-    op[key].multi = op.updateOne || op.replaceOne ? false : true;
-    return this.addToOperationsList(this, common.UPDATE, op[key]);
-  }
-
-  // Crud spec update format
-  if (op.updateOne || op.updateMany || op.replaceOne) {
-    var multi = op.updateOne || op.replaceOne ? false : true;
-    var operation = { q: op[key].filter, u: op[key].update || op[key].replacement, multi: multi };
-    operation.upsert = op[key].upsert ? true : false;
-    if (op.collation) operation.collation = op.collation;
-    if (op[key].arrayFilters) operation.arrayFilters = op[key].arrayFilters;
-    return this.addToOperationsList(this, common.UPDATE, operation);
-  }
-
-  // Remove operations
-  if (
-    op.removeOne ||
-    op.removeMany ||
-    (op.deleteOne && op.deleteOne.q) ||
-    (op.deleteMany && op.deleteMany.q)
-  ) {
-    op[key].limit = op.removeOne ? 1 : 0;
-    return this.addToOperationsList(this, common.REMOVE, op[key]);
-  }
-
-  // Crud spec delete operations, less efficient
-  if (op.deleteOne || op.deleteMany) {
-    var limit = op.deleteOne ? 1 : 0;
-    operation = { q: op[key].filter, limit: limit };
-    if (op.collation) operation.collation = op.collation;
-    return this.addToOperationsList(this, common.REMOVE, operation);
-  }
-
-  // Insert operations
-  if (op.insertOne && op.insertOne.document == null) {
-    if (forceServerObjectId !== true && op.insertOne._id == null) op.insertOne._id = new ObjectID();
-    return this.addToOperationsList(this, common.INSERT, op.insertOne);
-  } else if (op.insertOne && op.insertOne.document) {
-    if (forceServerObjectId !== true && op.insertOne.document._id == null)
-      op.insertOne.document._id = new ObjectID();
-    return this.addToOperationsList(this, common.INSERT, op.insertOne.document);
-  }
-
-  if (op.insertMany) {
-    for (var i = 0; i < op.insertMany.length; i++) {
-      if (forceServerObjectId !== true && op.insertMany[i]._id == null)
-        op.insertMany[i]._id = new ObjectID();
-      addToOperationsList(this, common.INSERT, op.insertMany[i]);
-    }
-
-    return;
-  }
-
-  // No valid type of operation
-  throw toError(
-    'bulkWrite only supports insertOne, insertMany, updateOne, updateMany, removeOne, removeMany, deleteOne, deleteMany'
-  );
+  common.Raw(op, this);
 };
 
 //
@@ -572,6 +406,13 @@ OrderedBulkOperation.prototype.execute = function(_writeConcern, options, callba
 
 FindOperatorsOrdered.prototype.addToOperationsList = addToOperationsList;
 OrderedBulkOperation.prototype.addToOperationsList = addToOperationsList;
+
+const isOrdered = function() {
+  return true;
+};
+
+FindOperatorsOrdered.prototype.isOrdered = isOrdered;
+OrderedBulkOperation.prototype.isOrdered = isOrdered;
 
 /**
  * Returns an unordered batch object
